requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.32.5
PyMuPDF==1.23.8
Pillow==10.0.0
numpy==1.26.4
opencv-python==4.8.1.78
scikit-learn==1.3.2
pandas==2.2.0
openpyxl==3.1.2
tqdm==4.66.1
ultralytics==8.3.175
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
openai==1.61.0
pyngrok==7.3.0
